Use specific route for login validation instead of systeminfo
Instead of using a call to /systeminfo which is resource intensive to test the authentication status, this would use a new backend route /auth/validate which has near zero resource cost.

The backend implementation is on [PR 6123](https://github.com/duplicati/duplicati/pull/6123) meaning the ngclient would only work on with that once its merged and released.