Support for new API on Websocket
This PR updates the code to support progress, task queue, and backup list updates served over websocket.

The client subscribes to the desired service and the server pushes updates when needed.

This gives a more smooth update experience, and avoids repeated calls through the REST API.

This update also supports the socket authentication where the token is not part of the query string, but instead passed as a separate initial authentication message.

Finally, this PR supports calculating an aggregated upload speed, if the server provides details on all transfers.

This fixes #209