Follow server backends
This PR makes the destinations reported in the UI follow the backends offered by the server.

This change supports cases where there is a mismatch between the FE/BE in terms of supported backends and available configurations.

If the user provides a non-existing backend in the target url, this is now mapped to a default configuration, so it can at least be shown. The edit capabilities are not great, but at least there is something.

In the case where there exists a UI configuration for a backend that is not supported by the backend, this backend is not offered by the UI, but can be accessed if the target url points to it.

This fixes #271
This fixes #272