Update to OAuth handling
This PR updates the default url used to prefer the server-supplied URL, and only use a built-in default if none is provided.

This also adds the OAuth v2 option that is needed for pCloud until we migrate away from GAE.

This also removes the token from the OAuth calls as the frame/window communication removes the need for it.