Implemented `waitForTaskToComplete` function
This PR introduces a `waitForTaskToComplete` method that returns an observable that can be awaited for completion.

If using websockets, this uses the socket data to check for task completion.

If not using sockets, or subscription is not available, it will revert to polling the tasks for completion.

This method is now used in the restore flow to wait for the partial database rebuild.