Reworked path handling
This PR introduces two new edit field types: Hostname and Bucketname.

These two fields validate the input as a hostname and bucket name respectively. They also prevent entering `/` into the text field as that trips up the parser because it splits host/path or bucket/path on the first `/`.

The destination config was updated to use appropriate types.

A few helper functions were added to avoid repeated parsing logic in the config file. Most configs can now call the `buildUrlFromFields` directly as the single combination step.

To get this working, the path field now has no leading slash in the path when editing, except when opting for root filesystem.