Added function to test if an item is a folder, based on the path.
Updated logic to not match startsWith for files, as that could cause (visual) mismatching:
- `/path/file1`
- `/path/file10`

Here, the second would previosly be counted as a child of the first.

Updated logic for toggling selected nodes to clear rules underneath. If a parent folder is selected, all children will be removed (as they are implicit) but excludes will remain.

If removing a path, all sub-paths are removed, including excludes.
This fixes #245